pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "active-inference-diffusion"
version = "0.1.0"
description = "Active Inference with Diffusion Models for Continuous Control"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Zahra Sheikhbahaee", email = "sheikhbahaee@gmail.com"},
]
keywords = ["active inference", "diffusion models", "reinforcement learning", "continuous control"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=1.12.0",
    "numpy>=1.21.0",
    "gymnasium>=0.28.0",
    "mujoco>=2.3.0",
    "tqdm>=4.65.0",
    "matplotlib>=3.5.0",
    "pyyaml>=6.0",
    "wandb>=0.13.0",
    "scipy>=1.7.0",
    "lz4>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[project.urls]
"Homepage" = "https://github.com/neuronphysics/active-inference-diffusion"
"Bug Reports" = "https://github.com/neuronphysics/active-inference-diffusion/issues"
"Source" = "https://github.com/neuronphysics/active-inference-diffusion"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]